move export queue db setup to separate file
